requirements.txt
numpy==1.14.1
opencv-python==3.4.0.12
svgpathtools==1.3.2
Pillow==5.1.0